Robots now display text when they die
Upon death, white text saying how many and of what type of minerals the
robot gathered appears over the robot, and fades out after some number
of frames. Right now the timeout is set to 30 frames, but we could
change that if we wanted. If a robot dies without collecting anything,
the text displayed is "Died in vain".